+++ src/infra/core/tests/tests/repos/test_metadata_chain_impl.rs
@@ -168,12 +168,13 @@ async fn test_append_prev_block_not_found() {
 
     assert_eq!(chain.resolve_ref(&BlockRef::Head).await.unwrap(), hash_1);
 
-    let block_2 = MetadataFactory::metadata_block(MetadataFactory::add_data().build())
-        .prev(
-            &Multihash::from_digest_sha3_256(b"does-not-exist"),
-            block_1_sequence_number,
-        )
-        .build();
+    let block_2 =
+        MetadataFactory::metadata_block(MetadataFactory::add_data().some_new_data().build())
+            .prev(
+                &Multihash::from_digest_sha3_256(b"does-not-exist"),
+                block_1_sequence_number,
+            )
+            .build();
 
     assert_matches!(
         chain.append(block_2, AppendOpts::default()).await,
@@ -366,10 +367,11 @@ async fn test_append_system_time_non_monotonic() {
 
     let hash_1 = chain.append(block_1, AppendOpts::default()).await.unwrap();
 
-    let block_2 = MetadataFactory::metadata_block(MetadataFactory::add_data().build())
-        .prev(&hash_1, block_1_sequence_number)
-        .system_time(Utc.with_ymd_and_hms(2000, 1, 1, 12, 0, 0).unwrap())
-        .build();
+    let block_2 =
+        MetadataFactory::metadata_block(MetadataFactory::add_data().some_new_data().build())
+            .prev(&hash_1, block_1_sequence_number)
+            .system_time(Utc.with_ymd_and_hms(2000, 1, 1, 12, 0, 0).unwrap())
+            .build();
 
     assert_matches!(
         chain.append(block_2, AppendOpts::default()).await,
@@ -1001,12 +1003,12 @@ async fn test_accept() {
 
     let mut always_stop_visitor = create_always_stop_visitor();
     let mut next_of_data_visitor = create_next_of_type_visitor(
-        MetadataBlockTypeFlags::DATA_BLOCK,
+        MetadataEventTypeFlags::DATA_BLOCK,
         NextOfTypeVisitorState::with_expected_visit_call_count(3),
     );
     let mut next_of_set_data_schema_visitor = create_next_of_type_visitor(
-        MetadataBlockTypeFlags::SET_DATA_SCHEMA,
-        NextOfTypeVisitorState::with_expected_visit_call_count(2),
+        MetadataEventTypeFlags::SET_DATA_SCHEMA,
+        NextOfTypeVisitorState::with_expected_visit_call_count(1),
     );
     let mut always_next_visitor = create_always_next_visitor_with_expected_visit_call_count(6);
 
@@ -1063,8 +1065,8 @@ async fn test_accept_stop_on_first_error() {
     let mut always_stop_visitor = create_always_stop_visitor();
     let mut always_next_visitor = create_always_next_visitor_with_expected_visit_call_count(2);
     let mut failed_on_type_visitor = create_failed_on_type_visitor_with_expected_visit_call_count(
-        MetadataBlockTypeFlags::SET_INFO,
-        2,
+        MetadataEventTypeFlags::SET_INFO,
+        1,
     );
 
     assert_matches!(
@@ -1075,7 +1077,7 @@ async fn test_accept_stop_on_first_error() {
                 &mut failed_on_type_visitor,
             ])
             .await,
-        Err(MockError::SomethingFailed)
+        Err(AcceptVisitorError::Visitor(MockError::SomethingFailed))
     );
 }
 
@@ -1092,12 +1094,6 @@ pub enum MockError {
     Internal(#[from] InternalError),
 }
 
-impl From<IterBlocksError> for MockError {
-    fn from(v: IterBlocksError) -> Self {
-        v.int_err().into()
-    }
-}
-
 ///////////////////////////////////////////////////////////////////////////////
 
 mockall::mock! {
@@ -1107,32 +1103,38 @@ mockall::mock! {
     impl MetadataChainVisitor for MetadataChainVisitor {
         type Error = MockError;
 
-        fn visit<'a>(&mut self, hashed_block_ref: HashedMetadataBlockRef<'a>) -> Result<Decision, MockError>;
+        fn initial_decision(&self) -> MetadataVisitorDecision;
+
+        fn visit<'a>(&mut self, hashed_block_ref: HashedMetadataBlockRef<'a>) -> Result<MetadataVisitorDecision, MockError>;
     }
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 
 fn create_failed_on_type_visitor_with_expected_visit_call_count(
-    fail_on_type_flags: MetadataBlockTypeFlags,
+    fail_on_type_flags: MetadataEventTypeFlags,
     visit_call_count: usize,
 ) -> MockMetadataChainVisitor {
-    let mut always_stop_visitor = MockMetadataChainVisitor::new();
+    let mut failed_on_type_visitor = MockMetadataChainVisitor::new();
 
-    always_stop_visitor
+    failed_on_type_visitor
+        .expect_initial_decision()
+        .times(1)
+        .returning(move || MetadataVisitorDecision::NextOfType(fail_on_type_flags));
+    failed_on_type_visitor
         .expect_visit()
         .times(visit_call_count)
         .returning(move |(_, block)| {
-            let block_flag = MetadataBlockTypeFlags::from(block);
+            let block_flag = MetadataEventTypeFlags::from(&block.event);
 
             if fail_on_type_flags.contains(block_flag) {
                 Err(MockError::SomethingFailed)
             } else {
-                Ok(Decision::NextOfType(fail_on_type_flags))
+                Ok(MetadataVisitorDecision::NextOfType(fail_on_type_flags))
             }
         });
 
-    always_stop_visitor
+    failed_on_type_visitor
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -1141,9 +1143,13 @@ fn create_always_stop_visitor() -> MockMetadataChainVisitor {
     let mut always_stop_visitor = MockMetadataChainVisitor::new();
 
     always_stop_visitor
-        .expect_visit()
+        .expect_initial_decision()
         .times(1)
-        .returning(|_| Ok(Decision::Stop));
+        .returning(|| MetadataVisitorDecision::Stop);
+    always_stop_visitor
+        .expect_visit()
+        .times(0)
+        .returning(|_| unreachable!());
 
     always_stop_visitor
 }
@@ -1156,9 +1162,13 @@ fn create_always_next_visitor_with_expected_visit_call_count(
     let mut always_next_visitor = MockMetadataChainVisitor::new();
 
     always_next_visitor
+        .expect_initial_decision()
+        .times(1)
+        .returning(|| MetadataVisitorDecision::Next);
+    always_next_visitor
         .expect_visit()
         .times(visit_call_count)
-        .returning(|_| Ok(Decision::Next));
+        .returning(|_| Ok(MetadataVisitorDecision::Next));
 
     always_next_visitor
 }
@@ -1180,13 +1190,17 @@ impl NextOfTypeVisitorState {
 }
 
 fn create_next_of_type_visitor(
-    flags: MetadataBlockTypeFlags,
+    flags: MetadataEventTypeFlags,
     state: NextOfTypeVisitorState,
 ) -> MockMetadataChainVisitor {
     let state = Arc::new(Mutex::new(state));
     let mut always_stop_visitor = MockMetadataChainVisitor::new();
     let expected_visit_call_count = state.lock().unwrap().expected_visit_call_count;
 
+    always_stop_visitor
+        .expect_initial_decision()
+        .times(1)
+        .returning(move || MetadataVisitorDecision::NextOfType(flags));
     always_stop_visitor
         .expect_visit()
         .times(expected_visit_call_count)
@@ -1196,9 +1210,9 @@ fn create_next_of_type_visitor(
             if state.visit_call_count != state.expected_visit_call_count {
                 state.visit_call_count += 1;
 
-                Ok(Decision::NextOfType(flags))
+                Ok(MetadataVisitorDecision::NextOfType(flags))
             } else {
-                Ok(Decision::Stop)
+                Ok(MetadataVisitorDecision::Stop)
             }
         });